Initial commit for telemetry retention demo
This import script uses elements of both example projects. The overall
structure is similar in terms of reading from the csv files and mapping the
elements to frames. This will initially handle segementation at the smallest
scope possible, to reduce complexity of the script.

The mapping of categorical labels to integers is simple and isn't persisted to
disk. This should be replaced with an additional key-value store like redis.